requirements.txt
torch
torchvision
onnx
onnxruntime-gpu
hydra-core
pandas
tqdm
pillow
